Update fork to last commit

+++ iped-parsers/iped-parsers-impl/src/main/java/dpf/sp/gpinf/indexer/parsers/jdbc/SQLite3DBParser.java
@@ -16,7 +16,6 @@ package dpf.sp.gpinf.indexer.parsers.jdbc;
  * limitations under the License.
  */
 
-import java.io.Closeable;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
@@ -39,6 +38,7 @@ import org.apache.tika.parser.ParseContext;
 import org.sqlite.SQLiteConfig;
 
 import dpf.sp.gpinf.indexer.parsers.util.DelegatingConnection;
+import dpf.sp.gpinf.indexer.util.IOUtil;
 import iped3.io.IItemBase;
 import iped3.search.IItemSearcher;
 import iped3.util.BasicProps;
@@ -49,6 +49,7 @@ import iped3.util.BasicProps;
  * This parser is internal only; it should not be registered in the services
  * file or configured in the TikaConfig xml file.
  */
+@SuppressWarnings("serial")
 public class SQLite3DBParser extends AbstractDBParser {
 
     protected static final String SQLITE_CLASS_NAME = "org.sqlite.JDBC"; //$NON-NLS-1$
@@ -75,30 +76,27 @@ public class SQLite3DBParser extends AbstractDBParser {
         TemporaryResources tmp = new TemporaryResources();
         try {
             File dbFile = TikaInputStream.get(stream, tmp).getFile();
-            File walTemp = exportWalLog(dbFile, context);
-            if(walTemp != null) {
-                tmp.addResource(new Closeable() {
-                    @Override
-                    public void close() {
-                        walTemp.delete();
-                        new File(dbFile.getAbsolutePath() + "-shm").delete();
-                    }
-                });
-            }
-            String connectionString = getConnectionString(dbFile);
-            
-            SQLiteConfig config = new SQLiteConfig();
+            boolean isTempDb = IOUtil.isTemporaryFile(dbFile);
+            if (isTempDb)
+                exportWalLog(dbFile, context);
 
-            // good habit, but effectively meaningless here
+            SQLiteConfig config = new SQLiteConfig();
             config.setReadOnly(true);
+
+            String connectionString = getConnectionString(dbFile);
             connection = config.createConnection(connectionString);
-            
+
             connection = new DelegatingConnection(connection) {
                 @Override
                 public void close() throws SQLException {
                     super.close();
                     try {
                         tmp.close();
+                        if (isTempDb) {
+                            // these files may be created by sqlite, even if wal was not exported
+                            new File(dbFile.getAbsolutePath() + "-wal").delete();
+                            new File(dbFile.getAbsolutePath() + "-shm").delete();
+                        }
                     } catch (IOException e) {
                         throw new SQLException(e);
                     }
@@ -110,21 +108,21 @@ public class SQLite3DBParser extends AbstractDBParser {
         }
         return connection;
     }
-    
+
     private File exportWalLog(File dbFile, ParseContext context) {
         IItemSearcher searcher = context.get(IItemSearcher.class);
-        if(searcher != null) {
+        if (searcher != null) {
             IItemBase dbItem = context.get(IItemBase.class);
-            if(dbItem != null) {
+            if (dbItem != null) {
                 String dbPath = dbItem.getPath();
                 String walQuery = BasicProps.PATH + ":\"" + searcher.escapeQuery(dbPath + "-wal") + "\"";
                 List<IItemBase> items = searcher.search(walQuery);
-                if(items.size() > 0) {
+                if (items.size() > 0) {
                     IItemBase wal = items.get(0);
                     File walTemp = new File(dbFile.getAbsolutePath() + "-wal");
-                    try(InputStream in = wal.getBufferedStream()){
+                    try (InputStream in = wal.getBufferedStream()) {
                         Files.copy(in, walTemp.toPath(), StandardCopyOption.REPLACE_EXISTING);
-                    }catch(IOException e) {
+                    } catch (IOException e) {
                         e.printStackTrace();
                     }
                     return walTemp;
@@ -133,11 +131,10 @@ public class SQLite3DBParser extends AbstractDBParser {
         }
         return null;
     }
-    
-    
-    
+
     @Override
-    protected String getConnectionString(InputStream stream, Metadata metadata, ParseContext context) throws IOException {
+    protected String getConnectionString(InputStream stream, Metadata metadata, ParseContext context)
+            throws IOException {
         throw new RuntimeException("Not Implemented"); //$NON-NLS-1$
     }
